Support giving xdg_popups keyboard focus.
This patch has basically disabled all the code previously written to lay
out windows. This is the result of a new architecture being needed in
the compositor. A popup view does not behave the same as a toplevel view
and thus we need a architecture which can better accomidate that.

This patch also migrates to the latest development master of wlroots.

Signed-off-by: Linus Johansen <linusen.a@gmail.com>